requirements.txt
unstructured
tiktoken
pinecone
pypdf
openai
langchain
langchain-community
pandas 
numpy
python-dotenv
langchain-pinecone
langchain-openai
networkx
openpyxl
pdfminer.six
PyMuPDF
tabula-py
pi-heif
unstructured-inference
pdf2image
pytesseract
pillow
huggingface_hub[hf_xet]
streamlit
pydantic
tqdm


